Variable distribution analysis
An R file has been added in which the analysis of the distribution of the variables has been performed. It includes graphs that were used to determine the amount of bias to be added to each variable.

This file is for visualization purposes only. All the work was later redone in Python in the file "data preparation.py"